Fin de la clase

+++ routes/users.js
@@ -1,41 +1,81 @@
 const router = require('express').Router();
 const User = require('../models/user.js');
+const bcrypt = require('bcrypt');
 
 router
   .get('/all', async (req, res) => {
-    console.log('GET /products/all');
+    console.log('GET /users/all');
     try {
-      const allProducts = await User.find();
-      res.status(200).send(allProducts);
+      const allUsers = await User.find();
+      res.status(200).send(allUsers);
     } catch (error) {
       res
         .status(400)
         .json({ error: true, message: error });
     }
   })
-  .get('/:id', async (req, res) => {
-    const { id } = req.params;
-    console.log('GET /products/' + id);
-    try {
-      const product = await User.findOne({
-        _id: id,
+  .post('/login', async (req, res) => {
+    const { body } = req;
+    console.log('POST /users/login');
+
+    const user = await User.findOne({
+      name: body.name,
+    });
+
+    const passwordOk = await bcrypt.compare(
+      body.password,
+      user.password
+    );
+
+    if (user && passwordOk) {
+      return res.status(200).json({
+        error: null,
+        message: 'User and password OK',
       });
-      res.status(200).json(product);
-    } catch (error) {
-      res.status(404).json({
+    } else {
+      return res.status(400).json({
         error: true,
-        message: error,
+        message: 'Credentials are WRONG',
       });
     }
   })
-  .post('/new', async (req, res) => {
-    console.log('POST /product/new');
+  .post('/register', async (req, res) => {
+    console.log('POST /users/register');
     const { body } = req;
+
+    const newUserNameExist = await User.findOne({
+      name: body.name,
+    });
+
+    const newUserMailExist = await User.findOne({
+      mail: body.mail,
+    });
+
+    // Cheque doble de previa existencia del usuario, en la API y en el Schema con unique
+    if (newUserNameExist || newUserMailExist) {
+      return res.status(400).json({
+        error: true,
+        message: 'User or email already exists',
+      });
+    }
+
+    // Aplico bcrypt
+    const salt = await bcrypt.genSalt(6);
+    const encryptedPassword = await bcrypt.hash(
+      body.password,
+      salt
+    );
+
     try {
-      const newProduct = new User(body);
-      await newProduct.save();
-      res.status(200).json(newProduct);
-      console.log('ADD id ' + newProduct._id);
+      const newUser = new User({
+        name: body.name,
+        mail: body.mail,
+        password: encryptedPassword,
+      });
+      await newUser.save();
+      newUser.password = body.password;
+      res.status(200).json(newUser);
+      console.log('ADD user ' + newUser.name);
     } catch (error) {
       console.log(error);
       res
@@ -43,17 +83,20 @@ router
         .json({ error: true, message: error });
     }
   })
-  .put('/update/:id', async (req, res) => {
-    const { id } = req.params;
+  .put('/update/:username', async (req, res) => {
+    const { username } = req.params;
     const { body } = req;
-    console.log('PUT/product/' + id);
+    console.log('PUT/users/update' + username);
     try {
-      const modProduct =
-        await User.findOneAndUpdate(id, body, {
+      const modUser = await User.findOneAndUpdate(
+        username,
+        body,
+        {
           useFindAndModify: false,
-        });
-      res.status(200).json(modProduct);
-      console.log('MOD id ' + modProduct._id);
+        }
+      );
+      res.status(200).json(modUser);
+      console.log('MOD user ' + modUser.name);
     } catch (error) {
       console.log(error);
       res.status(404).json({
@@ -62,23 +105,37 @@ router
       });
     }
   })
-  .delete('/delete/:id', async (req, res) => {
-    const { id } = req.params;
-    console.log('DELETE/product/' + id);
-    try {
-      const delProduct =
-        await User.findOneAndDelete({
-          _id: id,
+  .delete(
+    '/delete/:username',
+    async (req, res) => {
+      const { username } = req.params;
+      console.log('DELETE/users/' + username);
+
+      // chequeo previamente si el user es el super usuario para no borrarlo nunca
+      const SUPER_USER = 'admin';
+
+      if (username === SUPER_USER) {
+        return res.status(400).json({
+          error: true,
+          message: 'This user cannot be erased!',
         });
-      res.status(200).json(delProduct);
-      console.log('DEL id ' + delProduct._id);
-    } catch (error) {
-      console.log(error);
-      res.status(404).json({
-        error: true,
-        message: error,
-      });
+      }
+
+      try {
+        const delUser =
+          await User.findOneAndDelete({
+            name: username,
+          });
+        res.status(200).json(delUser);
+        console.log('DEL user ' + delUser.name);
+      } catch (error) {
+        console.log(error);
+        res.status(404).json({
+          error: true,
+          message: error,
+        });
+      }
     }
-  });
+  );
 
 module.exports = router;
